Yaml creation for as174 working

+++ communities/other.go
@@ -3,5 +3,5 @@ type Other struct {
 	What       string   `yaml:"what"`
 	Action string   `yaml:"action"`
 	From string   `yaml:"from"`
-	Community       int64   `yaml:"community"`
+	Community       int   `yaml:"community"`
 }
